+++ .babelrc
@@ -1,9 +1,9 @@
 {
-  "presets": ["es2015", "react", "stage-0"],
+  "presets": ["es2015", "react"],
   "plugins": [
-    [
-      "react-intl", {
-        "messagesDir": "./extracted"
-      }
-  ], "add-module-exports",]
+    ["react-intl", {
+        "messagesDir": "./build/messages/"
+        }],
+    "add-module-exports"
+  ]
 }
