+++ webpack/prod.config.js
@@ -1,16 +1,19 @@
 let webpack = require('webpack');
 let path = require('path');
+let StatsWriterPlugin = require('webpack-stats-plugin').StatsWriterPlugin;
+let Visualizer = require('webpack-visualizer-plugin');
 let ExtractTextPlugin = require('extract-text-webpack-plugin');
 
-
-module.exports = {
-    target: 'web',
+let webpackConfig = {
+    resolve: {
+        extensions: ['.js']
+    },
     entry: {
         main: [
             './client.js'
         ],
         vendor: [
-            'react', 'react-dom', 'react-intl', 'locale', 'react-hotkeys', 'react-list', 'react-responsive', 'react-custom-scrollbars', 'react-resize-aware', 'async', 'immutable', 'classnames', 'fluxible', 'fluxible-addons-react', 'fluxible-plugin-fetchr', 'fluxible-router', 'react-google-recaptcha', 'identicons-react', 'iso-639-1', 'lodash', 'cheerio', 'react-dnd', 'react-dnd-html5-backend', 'striptags', 'js-sha512', 'debug', 'md5', 'js-cookie', 'cookie', 'fumble', 'crypt'
+            'react', 'react-dom', 'react-hotkeys', 'react-list', 'react-responsive', 'react-custom-scrollbars', 'react-resize-aware', 'async', 'immutable', 'classnames', 'fluxible', 'fluxible-addons-react', 'fluxible-plugin-fetchr', 'fluxible-router', 'react-google-recaptcha', 'identicons-react', 'iso-639-1', 'lodash', 'cheerio', 'react-dnd', 'react-dnd-html5-backend', 'striptags', 'js-sha512', 'debug', 'md5', 'js-cookie', 'cookie', 'fumble', 'crypt'
         ]
     },
     output: {
@@ -19,17 +22,13 @@ module.exports = {
         filename: '[name].js'
     },
     module: {
-        loaders: [
+        rules: [
             {
-                test: /\.js$/,
-                exclude: /node_modules/,
+                test: /\.(js|jsx)$/,
+                exclude: /node_modules\/(?!identicons)/ ,
                 loader: 'babel-loader',
-                query: {
-                    presets: ['react', 'es2015', 'stage-0'],
-                    cacheDirectory: true
-                }
+                
             },
-
             {
                 test: /\.css$/,
                 loader: ExtractTextPlugin.extract({
@@ -42,8 +41,8 @@ module.exports = {
             { test   : /\.(ttf|eot|svg|woff(2)?)(\?[a-z0-9=&.]+)?$/, loader: 'url-loader?limit=100000'}
         ]
     },
-    devServer: {
-        contentBase: './assets'
+    node: {
+        setImmediate: false
     },
     plugins: [
         // css files from the extract-text-plugin loader
@@ -52,10 +51,37 @@ module.exports = {
             disable: false,
             allChunks: true
         }),
+
         new webpack.DefinePlugin({
             'process.env': {
-                NODE_ENV: JSON.stringify(process.env.NODE_ENV)
+                NODE_ENV: JSON.stringify('production'),
+                // Mainly used to require CSS files with webpack, which can happen only on browser
+                // Used as `if (process.env.BROWSER)...`
+                BROWSER: JSON.stringify(true),
             }
-        })
-    ]
+        }),
+        new webpack.optimize.UglifyJsPlugin({
+            sourceMap: true,
+            compress: {
+                warnings: false
+            }
+        }),
+        new webpack.optimize.CommonsChunkPlugin({
+            name: 'vendor',
+            minChunks: Infinity,
+            filename: '[name].bundle.js'
+        }),
+        // Write out stats file to build directory.
+        new StatsWriterPlugin({
+            filename: 'webpack.stats.json', // Default
+            fields: null,
+            transform: function (data) {
+                return JSON.stringify(data, null, 2);
+            }
+        }),
+        new Visualizer()
+    ],
+    devtool: 'source-map'
 };
+
+module.exports = webpackConfig;
